converted the Door actor to use blueprints by selecting the door, using the "Blueprints" menu -> "Convert selected actor to Blueprint Class". Then delete all the pre-populated events in the OpenDoor component.
Updated OpenDoor.h and OpenDoor.cpp to be compatible with and hookable by a Blueprint.

Then went back into the Door's blueprint -> OpenDoor component -> Right-click in blueprint area -> Add Event for OpenDoor -> Event Dispatchers -> Add On Open Request.

Click little white arrow on the "On Open Request (OpenDoor)" and connect it to a "SetActorRotation" event. Finally, modify the Z axis value for the "New Rotation" parameter of the SetActorRotation event.